Fix signed/unsigned mismatches, unaligned memory access, and add defensive checks

+++ src/activities/reader/FileSelectionActivity.cpp
@@ -206,7 +206,7 @@ void FileSelectionActivity::render() const {
 
   const auto pageStartIndex = selectorIndex / PAGE_ITEMS * PAGE_ITEMS;
   renderer.fillRect(0, 60 + (selectorIndex % PAGE_ITEMS) * THEME.itemHeight - 2, pageWidth - 1, THEME.itemHeight, THEME.selectionFillBlack);
-  for (int i = pageStartIndex; i < files.size() && i < pageStartIndex + PAGE_ITEMS; i++) {
+  for (size_t i = pageStartIndex; i < files.size() && i < pageStartIndex + PAGE_ITEMS; i++) {
     auto item = renderer.truncatedText(THEME.uiFontId, files[i].c_str(), renderer.getScreenWidth() - 40);
     const bool textColor = (i == selectorIndex) ? THEME.selectionTextBlack : THEME.primaryTextBlack;
     renderer.drawText(THEME.uiFontId, 20, 60 + (i % PAGE_ITEMS) * THEME.itemHeight, item.c_str(), textColor);
